Fix DoNotOptimize() GCC copy overhead (#1340)
The issue is on GCC DoNotOptimize() does full copy of an argument if it's not a
pointer and it slows down benchmark. If an argument is big enough
there is a memcpy() call for copy the agrument.
An argument object can be a big object so DoNotOptimize() could add sufficient
overhead and affect to benchmark results.

The cause is GCC behaviour with asm volatile constaints.
Looks like GCC trying to use r(register) constaint for all cases despite
object size. See https://gcc.gnu.org/bugzilla/show_bug.cgi?id=105519

The solution is the split DoNotOptimize() in two cases - value fits in register
and value doesn't fit in register. And use case specific asm constraint.
std::is_trivially_copyable trait is needed because "+r" constraint doesn't
work with non trivial copyable objects.

- Fix requires support C++11 feature std::is_trivially_copyable from GCC
  compiler. Feature is supported since GCC 5
- Fallback for GCC version < 5 still exists but it uses "m" constraint
  which means a little bit more overhead in some cases
- Add assembly tests for issued cases

fix #1340